+++ CHANGELOG.md
@@ -1,9 +1,167 @@
 # Toolbox Change Log
 
-## x.y.z
+## 1.0.8 - 2023-10-04
 
 ### Added
 
+- Nothing.
+
+### Changed
+
+- Nothing
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#60](https://github.com/mihaichris/atelier-toolbox/pull/60) Bump pillow from 9.4.0 to 10.0.1
+
+## 1.0.7 - 2023-09-21
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Nothing
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#59](https://github.com/mihaichris/atelier-toolbox/pull/59) Bump cryptography from 41.0.3 to 41.0.4
+
+## 1.0.6 - 2023-08-28
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Nothing
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#58](https://github.com/mihaichris/atelier-toolbox/pull/58) Bump pygments from 2.14.0 to 2.15.0
+- [#57](https://github.com/mihaichris/atelier-toolbox/pull/57) Bump cryptography from 41.0.2 to 41.0.3
+
+## 1.0.5 - 2023-08-28
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Nothing
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#56](https://github.com/mihaichris/atelier-toolbox/pull/56) Bump reportlab from 3.6.12 to 3.6.13
+
+## 1.0.4 - 2023-07-15
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Nothing
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#54](https://github.com/mihaichris/atelier-toolbox/pull/54) Bump cryptography from 41.0.0 to 41.0.2
+
+## 1.0.3 - 2023-06-08
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Security updates
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
+## 1.0.2 - 2023-02-08
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- [#52](https://github.com/mihaichris/atelier-toolbox/pull/52) Bump cryptography from 39.0.0 to 39.0.1
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
+## 1.0.1
+
+### Added
+
+- [#40](https://github.com/mihaichris/atelier-toolbox/pull/51) Add downloader cli subcommand
+- [#48](https://github.com/mihaichris/atelier-toolbox/pull/50) Add pre-commit install on make install command
+- [#14](https://github.com/mihaichris/atelier-toolbox/pull/49) Add hash subcommand
 - [#3](https://github.com/mihaichris/atelier-toolbox/pull/46) Generate requirements.txt on commit
 
 ### Changed